read.me
Initial Exploration
Learning the basics of exploratory data analysis using Python with Numpy,
Matplotlib, and Pandas.
What is EDA?
In Python, EDA is used for data visualization to generate meaningful patterns
and insights from data. It is also used for cleaning the data before analyzing it
to remove any abnormalities.
Based on the results of EDA, companies also make business decisions, such
as:
1- If EDA isn't done correctly, it can affect the further steps used in model
building.
2- If EDA is done properly, it can improve the efficiency of any model.
Import necessary packages
In Python, Numpy is a package that includes multidimensional array objects
as well as several derived objects.
Matplotlib is an amazing visualization library in Python for 2D plots of arrays.
Seaborn is on top of matplotlib for effective plot style
Pandas are used for data manipulation and analysis. So these are the core
libraries that are used for the EDA process.
Tasks:
1. find out the number of categorical features in our dataset.
2. We want to analyze the data to extract some insights
There are two major types of analysis data
● Univariate analysis
● Multivariate analysis
Univariate Analysis
● Univariate analysis is the simplest form of analyzing data.
● “Uni” means “one”, so in other words, data has only one variable.
● It doesn’t deal with causes or relationships and its major purpose is to
describe. It takes data, summarizes that data, and finds patterns in the data.
● Univariate Analysis can be done either on numerical or categorical
features.
3. Multivariate Analysis
In multivariate analysis, we try to find the relations between multiple variables.
Obviously, in real-life problems, variables can be any combination of numeric
or categorical variables. The combinations are:
● Numeric vs. Numeric
● Numeric vs. Categorical
● Categorical vs. Categorical
Another aspect of the variable combination is:
● Feature vs. Feature
● Feature vs. Target
4. Analysis of data with a variety of plots[Box plot,scatter plot,violin plot,bubble
plot,Histogram,Bar plot etc.,The Kernel Density Estimation